@@ -28,50 +28,94 @@
     "@mui/lab": "^5.0.0-alpha.53",
     "@mui/material": "^5.0.6",
     "@mui/styles": "^5.0.2",
+    "@types/d3": "^6.7.0",
     "@typescript-eslint/parser": "^5.20.0",
     "airbnb": "^0.0.2",
+    "babel-preset-react": "^6.24.1",
+    "bundle-loader": "^0.5.6",
     "chart.js": "^3.5.1",
-    "d3": "^7.0.1",
+    "d3": "^6.7.0",
     "date-fns": "^2.25.0",
+    "enzyme": "^3.8.0",
+    "enzyme-adapter-react-16": "^1.7.1",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-testing-library": "^5.3.1",
     "formik": "^2.1.5",
     "history": "^5.0.0",
+    "i18next": "^21.6.6",
+    "i18next-browser-languagedetector": "^6.1.2",
+    "i18next-http-backend": "^1.3.1",
+    "jest-enzyme": "^7.0.1",
     "jest-junit": "^13.2.0",
     "junit": "^1.4.9",
-    "next": "^12.0.3-canary.7",
-    "nextjs-mf": "^0.0.1",
+    "lodash": "^4.17.21",
+    "moment": "^2.29.1",
     "nprogress": "^0.2.0",
+    "process": "^0.11.10",
     "prop-types": "^15.7.2",
     "protobufjs": "^6.11.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-helmet": "^6.1.0",
+    "react-i18next": "^11.15.3",
     "react-perfect-scrollbar": "^1.5.8",
+    "source-map-loader": "^4.0.0",
+    "ts-loader": "^9.3.1",
     "yup": "^0.32.9"
   },
   "devDependencies": {
+    "@babel/preset-env": "^7.18.10",
+    "@babel/preset-typescript": "^7.18.6",
+    "@babel/runtime": "^7.17.9",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.4",
-    "@types/node": "^17.0.21",
+    "@types/enzyme": "^3.1.15",
+    "@types/enzyme-adapter-react-16": "^1.0.3",
+    "@types/jest": "^28.1.7",
+    "@types/node": "^18.7.6",
     "@types/react": "^17.0.40",
     "@types/react-dom": "^17.0.13",
+    "@types/react-helmet": "^6.1.5",
     "@typescript-eslint/eslint-plugin": "^5.10.0",
+    "babel-jest": "^28.1.3",
+    "babel-loader": "^8.2.5",
+    "css-loader": "^6.7.1",
     "eslint": "^8.7.0",
-    "eslint-config-next": "^12.0.8",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-react": "^7.28.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^4.0.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.12.3",
     "eslint-plugin-react-hooks": "^4.3.0",
+    "html-webpack-plugin": "^5.5.0",
     "husky": "^7.0.4",
     "jest": "^27.5.1",
+    "jest-fetch-mock": "^3.0.3",
+    "jest-junit": "^13.2.0",
+    "junit": "^1.4.9",
     "lint-staged": "^12.3.1",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.5.1",
     "prettier-plugin-packagejson": "^2.2.15",
     "prettier-plugin-sh": "^0.8.1",
-    "typescript": "^4.5.5",
+    "react-scripts": "^5.0.1",
+    "react-test-renderer": "^16.7.0",
+    "sass": "^1.49.4",
+    "sass-loader": "^12.6.0",
+    "style-loader": "^3.3.1",
+    "ts-jest": "^28.0.8",
+    "typescript": "^4.7.4",
+    "webpack": "^5.72.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.8.1",
     "yarn-deduplicate": "^3.1.0"
-  }
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
