@@ -1,18 +1,18 @@
+/* eslint-disable import/no-unresolved */
 import { useEffect, useState } from 'react';
 import { Card, CardContent, CardHeader, Container, Grid } from '@mui/material';
 
-import { Protocol } from 'src/models/protocol';
 import { MessageTopic } from 'src/models/message-topic';
 import { decodeJson } from 'src/libs/decoder';
 import { RfiQaPixelTable } from 'src/libs/rfi-qa-pixel-table';
 import { RfiDetailPlots } from 'src/libs/rfi-detail-plots';
 
-const WIDTH = 1200;
-const PROTOCOL = Protocol.JSON;
+import { PROTOCOL, WIDTH, WS_API_URL } from 'src/utils/constants';
+
 const MESSAGE_TOPIC = MessageTopic.RFI;
 const RFI_SUBTOPIC = 'xx-00-01';
-const RFI_API = `${process.env.NEXT_PUBLIC_WS_API}/${PROTOCOL}_${MESSAGE_TOPIC}`;
-const RFI_DETAILS_API = `${process.env.NEXT_PUBLIC_WS_API}/${PROTOCOL}_${MESSAGE_TOPIC}_${RFI_SUBTOPIC}`;
+const RFI_API = `${process.env.REACT_APP_WS_API}/${PROTOCOL}_${MESSAGE_TOPIC}`;
+const RFI_DETAILS_API = `${WS_API_URL}/${PROTOCOL}_${MESSAGE_TOPIC}_${RFI_SUBTOPIC}`;
 
 const Rfi = () => {
   const [socketStatus, setSocketStatus] = useState(Date().toLocaleString());
@@ -64,10 +64,7 @@ const Rfi = () => {
       <Grid container spacing={3}>
         <Grid item xs={12}>
           <Card sx={{ minWidth: WIDTH }}>
-            <CardHeader
-              title="RFI"
-              subheader={`Socket: ${socketStatus}`}
-            />
+            <CardHeader title="RFI" subheader={`Socket: ${socketStatus}`} />
 
             <CardContent sx={{ pt: '8px' }}>
               <div id="rfi-table-id" />
@@ -75,10 +72,7 @@ const Rfi = () => {
           </Card>
 
           <Card sx={{ minWidth: WIDTH }}>
-            <CardHeader
-              title={`RFI: ${RFI_SUBTOPIC}`}
-              subheader={`Socket: ${socketStatus}`}
-            />
+            <CardHeader title={`RFI: ${RFI_SUBTOPIC}`} subheader={`Socket: ${socketStatus}`} />
 
             <CardContent sx={{ pt: '8px' }}>
               <div id="rfi-details-id" />
