@@ -1,6 +1,6 @@
 import * as d3 from 'd3';
-import * as _ from 'lodash';
-import { removeLastDirectoryPartOf } from 'src/utils/common';
+// import * as _ from 'lodash';
+import { removeLastDirectoryPartOf } from '../utils/common';
 import { SpectrogramPlot } from './spectrogram-plot';
 
 interface Cell {
@@ -8,22 +8,35 @@ interface Cell {
   plot: SpectrogramPlot;
 }
 
-class SpectrogramPlotTable {
+export default class SpectrogramPlotTable {
   divId;
+
   width: number;
+
   height: number;
+
   cellHeight: number;
+
   cellWidth: number;
-  cellGap: number = 5;
+
+  cellGap = 5;
 
   table;
+
   cells: Cell[][];
+
   colNames;
+
   rowNames;
+
   numRows;
+
   numCols;
+
   data;
+
   len;
+
   unwrap = ({ baseline, polarisation }) => ({ baseline, polarisation });
 
   constructor(divId, width = 1200, height = 600, cellWidth = 200, cellHeight = 100) {
@@ -45,19 +58,19 @@ class SpectrogramPlotTable {
 
       // initialise 2d array of cells of a table
       this.cells = new Array(this.numRows);
-      for (let i = 0; i < this.numRows; i++) {
+      for (let i = 0; i < this.numRows; i += 1) {
         this.cells[i] = new Array(this.numCols);
       }
       // console.log("cells 1 = ", this.cells);
 
       // fill each cell with a data object
       let idx = 0;
-      for (let i = 0; i < this.numRows; i++) {
-        for (let j = 0; j < this.numCols; j++) {
+      for (let i = 0; i < this.numRows; i += 1) {
+        for (let j = 0; j < this.numCols; j += 1) {
           if (idx >= this.len) break;
           this.cells[i][j] = {} as Cell;
           this.cells[i][j].metadata = { ...this.unwrap(data[idx]), idx };
-          idx++;
+          idx += 1;
         }
       }
       // console.log("cells 2 = ", this.cells);
@@ -66,8 +79,8 @@ class SpectrogramPlotTable {
     }
 
     let idx = 0;
-    for (let i = 0; i < this.numRows; i++) {
-      for (let j = 0; j < this.numCols; j++) {
+    for (let i = 0; i < this.numRows; i += 1) {
+      for (let j = 0; j < this.numCols; j += 1) {
         if (idx >= this.len) break;
         // console.log("cells 3 = ", this.cells[i][j], idx);
 
@@ -78,21 +91,16 @@ class SpectrogramPlotTable {
         }
 
         this.cells[i][j].plot?.draw(this.data[idx].phase);
-        idx++;
+        idx += 1;
       }
     }
   }
 
   drawTable() {
     // clear/remove existing table
-    d3.select('#' + this.divId)
-      .selectAll('table')
-      .remove();
+    d3.select(`#${this.divId}`).selectAll('table').remove();
 
-    this.table = d3
-      .select('#' + this.divId)
-      .append('table')
-      .style('class', 'table');
+    this.table = d3.select(`#${this.divId}`).append('table').style('class', 'table');
 
     const tablebody = this.table.append('tbody');
     const rows = tablebody.selectAll('tr').data(this.cells).enter().append('tr');
@@ -101,15 +109,15 @@ class SpectrogramPlotTable {
     rows
       .selectAll('td')
       // each row has data associated; we get it and enter it for the cells.
-      .data((d, i) => {
+      .data((d) => {
         return d;
       })
       .enter()
       .append('td')
-      .text((d, i) => {
+      .text((d) => {
         return this.getName(d?.metadata);
       })
-      .on('click', (d, i) => {
+      .on('click', (i) => {
         if (i?.metadata?.idx !== undefined) {
           window.open(
             `${removeLastDirectoryPartOf(window.location.href)}/spectrogram/?idx=${
@@ -119,24 +127,21 @@ class SpectrogramPlotTable {
         }
       })
       .append('canvas')
-      .attr('id', (d, i) => {
+      .attr('id', (d) => {
         const id = this.getId(d?.metadata);
         if (id) return id;
+        return null;
       })
       .attr('style', 'canvas')
       .attr('width', this.cellWidth)
       .attr('height', this.cellHeight);
-
-    return;
   }
 
-  private getName(d: any) {
+  private getName(d) {
     return d?.baseline && d?.polarisation ? `${d?.baseline}-${d?.polarisation}` : '-';
   }
 
-  private getId(d: any) {
+  private getId(d) {
     return d?.baseline && d?.polarisation ? `canvas-${d?.baseline}-${d?.polarisation}` : undefined;
   }
 }
-
-export default SpectrogramPlotTable;
