@@ -2,14 +2,20 @@ import * as d3 from 'd3';
 
 export class SpectrumPlotSvg {
   width: number;
+
   height: number;
+
   margin = { top: 10, right: 40, bottom: 60, left: 50 };
 
-  svg: any;
-  xLabel: string = 'Frequency (MHz)';
-  yLabel: string = 'Power (dB)';
-  xScale: any;
-  yScale: any;
+  svg;
+
+  xLabel = 'Frequency (MHz)';
+
+  yLabel = 'Power (dB)';
+
+  xScale;
+
+  yScale;
 
   constructor(selector: string, width = 1200, height = 600) {
     this.width = width;
@@ -25,8 +31,8 @@ export class SpectrumPlotSvg {
       .attr('transform', `translate(${this.margin.left},${this.margin.top})`);
   }
 
-  public draw(data: any) {
-    //console.log("SpectrumPlot:draw: data = ", data);
+  public draw(data) {
+    // console.log("SpectrumPlot:draw: data = ", data);
     this.svg.selectAll('text').remove();
     this.svg.selectAll('.tick').remove();
     this.svg.selectAll('path').remove();
