@@ -1,34 +1,36 @@
-export const mockSpectrogramsData = {
+const mockSpectrogramsData = {
   spectrogram: [
     {
       timestamp: '2022-03-15 10:23:17',
       baseline: '0',
       polarisation: 'XY',
-      phase: [148, 202, 284, 205, 185, 246, 108, 33, 214, 38],
+      phase: [148, 202, 284, 205, 185, 246, 108, 33, 214, 38]
     },
     {
       timestamp: '2022-03-15 10:23:17',
       baseline: '1',
       polarisation: 'YY',
-      phase: [153, 350, 213, 179, 205, 20, 238, 102, 168, 345],
+      phase: [153, 350, 213, 179, 205, 20, 238, 102, 168, 345]
     },
     {
       timestamp: '2022-03-15 10:23:17',
       baseline: '2',
       polarisation: 'XX',
-      phase: [85, 193, 205, 239, 141, 30, 159, 320, 112, 113],
+      phase: [85, 193, 205, 239, 141, 30, 159, 320, 112, 113]
     },
     {
       timestamp: '2022-03-15 10:23:17',
       baseline: '3',
       polarisation: 'YY',
-      phase: [210, 64, 280, 99, 62, 37, 178, 224, 228, 193],
+      phase: [210, 64, 280, 99, 62, 37, 178, 224, 228, 193]
     },
     {
       timestamp: '2022-03-15 10:23:17',
       baseline: '4',
       polarisation: 'YX',
-      phase: [12, 142, 275, 237, 192, 14, 316, 321, 125, 70],
-    },
-  ],
+      phase: [12, 142, 275, 237, 192, 14, 316, 321, 125, 70]
+    }
+  ]
 };
+
+export default mockSpectrogramsData;
