@@ -1,22 +1,23 @@
 //
 // This data structure is out of date!
 //
-export const mockSpectrumData = {
+const mockSpectrumData = {
   timestamp: '2022-03-14 13:58:18',
   xMax: 10,
   yMax: 10,
   channels: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9],
   power: [
     0.8999999761581421, 1.600000023841858, 1.100000023841858, 5, 0.5, 1.899999976158142,
-    6.900000095367432, 2, 7, 5,
+    6.900000095367432, 2, 7, 5
   ],
   sdL: [
     0.6499999761581421, 0.5099999904632568, 0.27000001072883606, 0.46000000834465027,
     0.550000011920929, 0.7799999713897705, 0.3199999928474426, 0.5899999737739563,
-    0.36000001430511475, 0.5799999833106995,
+    0.36000001430511475, 0.5799999833106995
   ],
   sdU: [
     0.27000001072883606, 0.3100000023841858, 0.6700000166893005, 0.699999988079071,
-    0.23999999463558197, 0.5699999928474426, 0.4000000059604645, 0.7099999785423279, 0.5, 0.75,
-  ],
+    0.23999999463558197, 0.5699999928474426, 0.4000000059604645, 0.7099999785423279, 0.5, 0.75
+  ]
 };
+export default mockSpectrumData;
