@@ -2,5 +2,6 @@ export enum MessageTopic {
   SPECTRUM = 'spectrum',
   SPECTROGRAMS = 'spectrograms',
   SPECTROGRAM = 'spectrogram',
-  RFI = 'rfi',
+  RFI = 'rfi'
 }
+
