@@ -1,5 +1,5 @@
 export function removeLastDirectoryPartOf(url: string) {
-  let arr = url.split('/');
+  const arr = url.split('/');
   arr.pop();
   return arr.join('/');
 }
